Adding a data integrity system. Should allow for easy integrity management.

+++ generators/app/templates/MainSolutionTemplate/src/MainSolutionTemplate.Core/BusinessLogic/Components/BaseManager.cs
@@ -3,16 +3,17 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
-using System.Runtime.InteropServices;
 using System.Threading.Tasks;
+using log4net;
 using MainSolutionTemplate.Core.BusinessLogic.Components.Interfaces;
+using MainSolutionTemplate.Core.DataIntegrity;
 using MainSolutionTemplate.Core.MessageUtil;
 using MainSolutionTemplate.Core.MessageUtil.Models;
 using MainSolutionTemplate.Dal.Models;
 using MainSolutionTemplate.Dal.Models.Enums;
 using MainSolutionTemplate.Dal.Persistance;
 using MainSolutionTemplate.Dal.Validation;
-using log4net;
+using MainSolutionTemplate.Utilities.Helpers;
 
 namespace MainSolutionTemplate.Core.BusinessLogic.Components
 {
@@ -21,12 +22,14 @@ namespace MainSolutionTemplate.Core.BusinessLogic.Components
         protected readonly IGeneralUnitOfWork _generalUnitOfWork;
         protected readonly IMessenger _messenger;
         protected readonly IValidatorFactory _validationFactory;
+        protected IDataIntegrityManager _dataIntegrityManager;
 
         protected BaseManager(BaseManagerArguments baseManagerArguments)
         {
             _generalUnitOfWork = baseManagerArguments.GeneralUnitOfWork;
             _messenger = baseManagerArguments.Messenger;
             _validationFactory = baseManagerArguments.ValidationFactory;
+            _dataIntegrityManager = baseManagerArguments.DataIntegrityManager;
         }
     }
 
@@ -44,7 +47,6 @@ namespace MainSolutionTemplate.Core.BusinessLogic.Components
 
         #region IBaseManager<T> Members
 
-      
         public Task<List<T>> Get(Expression<Func<T, bool>> filter)
         {
             return Repository.Find(filter);
@@ -60,22 +62,20 @@ namespace MainSolutionTemplate.Core.BusinessLogic.Components
             return Repository.FindOne(x => x.Id == id);
         }
 
-        public virtual async Task<T> Save(T entity)
-        {
-            T projectFound =  await Get(entity.Id);
-            if (projectFound == null)
-            {
-                return await Insert(entity);
-            }
-            return await Update(entity); 
-        }
-
         public virtual async Task<T> Delete(Guid id)
         {
             T project = await Get(id);
             if (project != null)
             {
                 _log.Info(string.Format("Remove {1} [{0}]", project, _name));
+                long count = await _dataIntegrityManager.GetReferenceCount(project);
+                if (count > 0)
+                {
+                    throw new Exception(
+                        string.Format(
+                            "Could not remove {0} [{1}]. It is currently referenced in {2} other data object.",
+                            typeof (T).Name.UnderScoreAndCamelCaseToHumanReadable(), project, count));
+                }
                 await Repository.Remove(x => x.Id == id);
                 _messenger.Send(new DalUpdateMessage<T>(project, UpdateTypes.Removed));
             }
@@ -87,28 +87,47 @@ namespace MainSolutionTemplate.Core.BusinessLogic.Components
             return Repository.Query();
         }
 
-        #endregion
-
-        protected async Task<T> Update(T entity)
+        public async Task<T> Update(T entity)
         {
             DefaultModelNormalize(entity);
             await Validate(entity);
             _log.Info(string.Format("Update {1} [{0}]", entity, _name));
             T update = await Repository.Update(x => x.Id == entity.Id, entity);
+            _dataIntegrityManager.UpdateAllReferences(update).ContinueWithNoWait(LogUpdate);
             _messenger.Send(new DalUpdateMessage<T>(entity, UpdateTypes.Updated));
             return update;
         }
 
         
 
-        protected async Task<T> Insert(T entity)
+        public async Task<T> Insert(T entity)
         {
             DefaultModelNormalize(entity);
             await Validate(entity);
             _log.Info(string.Format("Adding {1} [{0}]", entity, _name));
-            var insert = await Repository.Add(entity);
+            T insert = await Repository.Add(entity);
             _messenger.Send(new DalUpdateMessage<T>(entity, UpdateTypes.Inserted));
             return insert;
+        }
+
+        #endregion
+
+        public virtual async Task<T> Save(T entity)
+        {
+            T projectFound = await Get(entity.Id);
+            if (projectFound == null)
+            {
+                return await Insert(entity);
+            }
+            return await Update(entity);
+        }
+
+        private void LogUpdate(Task<long> obj)
+        {
+            if (obj.Result > 0)
+            {
+                _log.Info("{0} referenced items have been updated.");
+            }
         }
 
         protected virtual void DefaultModelNormalize(T user)
